Refactored RulesTable.tsx - separated logic to a new file - useRulesTableLogic.tsx, in RulesTable.tsx left only the view.
